fix: handle BigInt inviteNonce at JS serialization boundaries
After schema.prisma changed inviteNonce to BigInt?, Prisma returns the
value as a JavaScript bigint, which cannot be JSON-serialised (breaks
JWT sign()) and does not match what Prisma expects in a where clause
when the value arrives as a plain number from a decoded JWT.

- schema/index.js: inviteNonce: BigInt on Round GraphQL type (was Int,
  which overflows for 13-digit Date.now() values)
- queries/round.ts + queries/group.ts: Number(x.inviteNonce) before
  sign() so the JWT payload carries a serialisable number
- mutations/round.ts: BigInt(inviteNonce) in both findFirst where
  clauses so Prisma receives the correct type on link redemption

+++ ui/server/graphql/schema/index.js
@@ -505,7 +505,7 @@ const schema = gql`
     discourseCategoryId: Int
     tags: [Tag!]
     bucketStatusCount: BucketStatusCount
-    inviteNonce: Int
+    inviteNonce: BigInt
     updatedAt: Date
     distributedAmount: Int
     publishedBucketCount: Int
